fix: fix bugs on password resetting functionality - usecase/password_reset/code_validate_command.go

+++ backend/AAiT-backend-group-13/usecase/password_reset/code_validate_command.go
@@ -5,12 +5,12 @@ import ()
 // ValidateCodeCommand represents the command to validate a reset code.
 // It includes the reset code and the user ID associated with it.
 type ValidateCodeCommand struct {
-	code  int64
+	code  int
 	email string
 }
 
 // NewValidateCodeCommand creates a new instance of ValidateCodeCommand.
-func NewValidateCodeCommand(code int64, email string) *ValidateCodeCommand {
+func NewValidateCodeCommand(code int, email string) *ValidateCodeCommand {
 	return &ValidateCodeCommand{
 		code:  code,
 		email: email,
